requirements.txt
Django==1.11.1
djangorestframework==3.6.3
Pillow==4.1.1